Move fixtures loading in a service
See #392
