+++ src/Resources/config/functional_test.xml
@@ -23,5 +23,12 @@
             <argument>%liip_functional_test.query.max_query_count%</argument>
             <argument type="service" id="annotation_reader" />
         </service>
+
+        <service id="liip_functional_test.fixtures_loader" class="Liip\FunctionalTestBundle\Service\FixturesLoader" public="true">
+            <!--<argument type="service" id="doctrine"/>-->
+            <!--<argument>%liip_functional_test.cache_sqlite_db%</argument>-->
+            <!--<argument>%kernel.cache_dir%</argument>-->
+            <argument type="service" id="service_container"/>
+        </service>
     </services>
 </container>
